refactor: move history utils to separate store
