+++ frontend/src/stores/slide.js
@@ -1,7 +1,8 @@
 import { ref, computed, reactive } from 'vue'
-import { ignoreUpdates, slidesLength, presentationId, templateList, readonlyMode, presentationTheme, showLayoutsView } from '@/stores/presentation'
+import { slidesLength, presentationId, templateList, readonlyMode, presentationTheme, showLayoutsView } from '@/stores/presentation'
 import { resetFocus } from '@/stores/element'
 import { saveChanges, isDirty } from '@/stores/saving'
+import { ignoreUpdates } from '@/stores/history'
 import { generateUniqueId, cloneObj } from '@/utils/helpers'
 import { router } from '@/router'
